+++ docs/changelog.md
@@ -18,6 +18,7 @@ Audio driver changes:
     * `vol_l` and `vol_r` are now calculated after the bytecode has been processed.
  * `return_from_subroutine` bytecode instruction disables the channel if it is not in a subroutine.
  * Optimised `skip_last_loop` bytecode instructions
+ * Added sound effect dropout behaviour
 
 IO Command changes:
  * Removed `SET_ENABLED_CHANNELS` IO command, replaced with `SET_MUSIC_CHANNELS`.